Fix P1 high-priority bugs: scalar precision, DivBackward broadcast, thread-local RNG, CUDA error handling
- Pass double directly to full() in scalar arithmetic ops instead of
  truncating to float, preserving Float64 precision
- Add broadcast gradient reduction to DivBackward matching MulBackward
  pattern — save input shapes in forward and operator/, reduce gradients
  in both backward() and backward_with_variables()
- Make global RNG thread_local in creation.cpp so concurrent rand/randn
  calls don't race on shared state
- Replace per-file CUDA_CHECK macros (using exit(EXIT_FAILURE)) with
  centralized cuda_error.hpp (using throw) in 5 kernel files
- Fix backward_with_variables in ReshapeBackward, PermuteBackward,
  TransposeBackward, SqueezeBackward to use Variable-level operations
  for correct higher-order gradient flow

+++ src/ops/math.cpp
@@ -116,7 +116,7 @@ auto add(const Tensor& a, double scalar) -> Tensor {
     if (!a.impl()) {
         throw std::runtime_error("Cannot add to uninitialized tensor");
     }
-    auto scalar_tensor = full({1}, static_cast<float>(scalar), a.dtype(), a.device());
+    auto scalar_tensor = full({1}, scalar, a.dtype(), a.device());
     std::vector<Tensor> inputs = {a.is_contiguous() ? a : a.contiguous(), scalar_tensor};
     return dispatch<OpId::Add>(inputs)[0];
 }
@@ -125,7 +125,7 @@ auto sub(const Tensor& a, double scalar) -> Tensor {
     if (!a.impl()) {
         throw std::runtime_error("Cannot subtract from uninitialized tensor");
     }
-    auto scalar_tensor = full({1}, static_cast<float>(scalar), a.dtype(), a.device());
+    auto scalar_tensor = full({1}, scalar, a.dtype(), a.device());
     std::vector<Tensor> inputs = {a.is_contiguous() ? a : a.contiguous(), scalar_tensor};
     return dispatch<OpId::Sub>(inputs)[0];
 }
@@ -134,7 +134,7 @@ auto mul(const Tensor& a, double scalar) -> Tensor {
     if (!a.impl()) {
         throw std::runtime_error("Cannot multiply uninitialized tensor");
     }
-    auto scalar_tensor = full({1}, static_cast<float>(scalar), a.dtype(), a.device());
+    auto scalar_tensor = full({1}, scalar, a.dtype(), a.device());
     std::vector<Tensor> inputs = {a.is_contiguous() ? a : a.contiguous(), scalar_tensor};
     return dispatch<OpId::Mul>(inputs)[0];
 }
@@ -143,7 +143,7 @@ auto div(const Tensor& a, double scalar) -> Tensor {
     if (!a.impl()) {
         throw std::runtime_error("Cannot divide uninitialized tensor");
     }
-    auto scalar_tensor = full({1}, static_cast<float>(scalar), a.dtype(), a.device());
+    auto scalar_tensor = full({1}, scalar, a.dtype(), a.device());
     std::vector<Tensor> inputs = {a.is_contiguous() ? a : a.contiguous(), scalar_tensor};
     return dispatch<OpId::Div>(inputs)[0];
 }
